+++ qlbm_mcgill/ibm_qpu.py
@@ -16,9 +16,10 @@ class IBM_QPU_Runner(Runner):
     stm_lattice: SpaceTimeLattice
     lattice: CollisionlessLattice | Lattice
     job_id: str
-    dims: tuple
+    dims: list | tuple
     service: QiskitRuntimeService
     backend: IBMBackend
+    transpiled_circuits: list[QuantumCircuit]
     label: str
 
     def __init__(
@@ -58,7 +59,7 @@ class IBM_QPU_Runner(Runner):
         step_qcs = [StepCircuit(self.lattice, i, collision=False, init_cond=init_cond).circuit for i in range(steps+1)]
 
         pass_manager = generate_preset_pass_manager(optimization_level=1, backend=self.backend)
-        qcs = [pass_manager.run(qc) for qc in step_qcs]
+        self.transpiled_circuits = [pass_manager.run(qc) for qc in step_qcs]
 
         options = SamplerOptions()
         options.dynamical_decoupling.enable = True
@@ -70,7 +71,7 @@ class IBM_QPU_Runner(Runner):
         print("done.")
         print(f"Sending {steps} step job to {self.backend} with {shots} shots per time-step... ", end="")
 
-        job = sampler.run(qcs, shots=shots)
+        job = sampler.run(self.transpiled_circuits, shots=shots)
         self.job_id = job.job_id()
 
         print("done.")
@@ -85,6 +86,7 @@ class IBM_QPU_Runner(Runner):
             steps: int, 
             shots: int,
             readout_error_mitigation: bool = False,
+            iterative_bayesian_unfolding: bool = False
         ) -> str:
         """
         Takes the counts of the measurement data from the IBM QPU and turns it into a PyVista simulation.
@@ -97,8 +99,14 @@ class IBM_QPU_Runner(Runner):
         results = job.result()
         raw_counts = [list(results[i].data.values())[0].get_counts() for i in range(steps+1)]
 
-        error_mitigator = ErrorMitigator(self.lattice, readout_error_mitigation=readout_error_mitigation)
-        counts, self.label = error_mitigator.readout_error(shots, raw_counts, self.backend)
+        error_mitigator = ErrorMitigator(
+            self.lattice, 
+            self.backend,
+            readout_error_mitigation=readout_error_mitigation,
+            iterative_bayesian_unfolding=iterative_bayesian_unfolding
+        )
+
+        counts, self.label = error_mitigator.mitigate(self.transpiled_circuits, shots, raw_counts)
 
         rmdir_rf(f"ibm-qpu-output\\{self.label}")
         create_directory_and_parents(f"ibm-qpu-output\\{self.label}")
@@ -122,6 +130,7 @@ class IBM_QPU_Runner(Runner):
             steps: int, 
             shots: int = DEFAULT_SHOTS, 
             readout_error_mitigation: bool = False,
+            iterative_bayesian_unfolding: bool = False,
             init_cond: None | QuantumCircuit = None
         ) -> str:
         """
@@ -140,5 +149,9 @@ class IBM_QPU_Runner(Runner):
             print(f"Time elapsed: {int(diff/60)} minute(s) and {diff % 60} second(s).", end="")
         print(f"\nData received. Workload: {int(job.usage())} seconds.")
         time.sleep(2) # make them wait for it.
-        vis = self.visualize(steps, shots=shots, readout_error_mitigation=readout_error_mitigation)
+        vis = self.visualize(
+            steps, shots=shots, 
+            readout_error_mitigation=readout_error_mitigation,
+            iterative_bayesian_unfolding=iterative_bayesian_unfolding
+        )
         return vis
